Remove permission system support
The API comes with a permission query. Luckily nowadays permission systems are able to fully control that one.
It is not required (anymore) to depend on each existing permission plugin. While the most common one is not supported by this plugin anyway.
Thus removed all (already commented out) permission code and instead called the provided API.