Use ChatColor where possible
There is no need to colorize over and over again, but only once when the text is created.
Yes the readability decreases, but number-codes are not perfect either.
This way the color-striping can be done by Bukkit though.
Also cleaned up the help-text method.
Also removed redundant messages where possible.